test: add STOW-RS success integration tests
- Add 25 tests for successful DICOM storage via STOW-RS
- Test POST vs PUT behavior (duplicate handling, upsert)
- Test multipart/related parsing with various boundaries
- Verify response format matches DICOM JSON PS3.18 spec
- Verify database persistence and queryability via QIDO-RS
- Verify change feed entry creation

Coverage improvements:
- app/routers/dicomweb.py: 18.71% → 59.15% (+40.44%)
- app/services/multipart.py: 13.89% → 88.89% (+75.00%)

Bug fix:
- Change ChangeFeedEntry.sequence from BigInteger to Integer
  for SQLite autoincrement compatibility

Test infrastructure:
- Set DICOM_STORAGE_DIR environment variable before module imports
- Monkeypatch storage directory for isolated test execution

All 25 tests pass successfully.

+++ app/models/dicom.py
@@ -93,7 +93,7 @@ class ChangeFeedEntry(Base):
 
     __tablename__ = "change_feed"
 
-    sequence = Column(BigInteger, primary_key=True, autoincrement=True)
+    sequence = Column(Integer, primary_key=True, autoincrement=True)
     study_instance_uid = Column(String(128), nullable=False)
     series_instance_uid = Column(String(128), nullable=False)
     sop_instance_uid = Column(String(128), nullable=False)
